moved more functionality to agentmanager
